+++ src/main/kotlin/io/acari/doki/settings/actors/ThemeStatusBarActor.kt
@@ -1,0 +1,11 @@
+package io.acari.doki.settings.actors
+
+import io.acari.doki.config.ThemeConfig
+
+object ThemeStatusBarActor {
+  fun applyConfig(showThemeStatusBar: Boolean) {
+    if(ThemeConfig.instance.showThemeStatusBar != showThemeStatusBar) {
+      ThemeConfig.instance.showThemeStatusBar = showThemeStatusBar
+    }
+  }
+}