+++ changelog/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 ---
+# 5.0.2
+
+- Brought back configurable the theme name in the lower right hand of the IDE.
+    - Component no longer takes you to the settings page.
+
 # 5.0.1
 
 - Opt in to Theme transition animation, as that configuration sticks around after removing the plugin.